clean code：混用有符号数和无符号数
# message auto-generated for no-merge-commit merge:
merge clean-code into master

clean code：混用有符号数和无符号数

Created-by: msprofDeveloper

Match-id-643468a5badbedfd0fb0a4b80a8501a677563a96

+++ analysis/csrc/application/credential/id_pool.h
@@ -27,7 +27,7 @@ public:
     void Clear();
     uint64_t GetId(const CorrelationTuple& key);
     uint64_t GetUint64Id(const std::string& key);
-    uint64_t GetUint32Id(const std::string& key);
+    uint32_t GetUint32Id(const std::string& key);
     std::unordered_map<std::string, uint64_t>& GetAllUint64Ids();
 private:
     uint64_t correlationIndex_ = 0;
